Mobile Hooks applied. SetUpMobile split and modified.
