+++ src/test/java/org/Globant/pageObjects/mobile/SetUpMobile.java
@@ -1,13 +1,6 @@
 package org.Globant.pageObjects.mobile;
 
-import io.appium.java_client.android.AndroidDriver;
-import io.appium.java_client.android.AndroidElement;
-import org.openqa.selenium.remote.DesiredCapabilities;
-import org.Globant.utils.configuration.ConfigCapabilities;
-import org.Globant.configuration.mobile.UniqueMobileDriver;
-
-import java.net.MalformedURLException;
-import java.net.URL;
+import static org.Globant.steps.mobile.MobileHooks.getDriver;
 
 /**
  * Base class for Test classes.
@@ -17,49 +10,10 @@ import java.net.URL;
 public class SetUpMobile {
 
     protected static TutorialScreen tutorialScreen;
-    public static UniqueMobileDriver driver;
 
 
     public static void setUpStartApp() {
         tutorialScreen = new TutorialScreen(getDriver());
-    }
-
-    /**
-     * SetUp before to run suite of test.
-     *
-     * @author Arley.Bolivar
-     */
-
-    public static void environmentSetUp() {
-        DesiredCapabilities capabilities = new DesiredCapabilities();
-        ConfigCapabilities.deviceSetUp(capabilities);
-        ConfigCapabilities.applicationSetUp(capabilities);
-        try {
-            driver = new UniqueMobileDriver(new AndroidDriver<>(new URL("http://127.0.0.1:4723/wd/hub"), capabilities));
-        } catch (MalformedURLException exception) {
-            exception.printStackTrace();
-        }
-        setUpStartApp();
-    }
-
-    /**
-     * Close the application after completing the test.
-     *
-     * @author Arley.Bolivar
-     */
-
-    public static void mobileApplicationEnd() {
-        driver.getDriver().quit();
-    }
-
-    /**
-     * return the driver.
-     *
-     * @return driver
-     * @author Arley.Bolivar
-     */
-    public static AndroidDriver<AndroidElement> getDriver() {
-        return driver.getDriver();
     }
 
     /**
